pyproject.toml
[project]
name = "specification-assistant"
version = "0.1.0"
description = "PDF text extraction service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pdfplumber>=0.11.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "jinja2>=3.1.0",
    "langchain-openai>=1.0.1",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    # Database and authentication
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.3.0",  # Pin bcrypt for passlib compatibility
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_reader"]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
